Test new structure, incl. GoalRowContainer
Now that GoalList and GoalsTable have the correct relationship, tests are req.

Additional work was done to wrap the GoalRow in a container in order to inject
the action creators, rather than passing them all the way down from the top in
the GoalsTable.

Misc. tests were added, unrelated to this functionality, I just happened to see
they were missing.