Abstract DebounceComponent
Previously, debouncing was done through a bespoke input, which was
unneccessarily restrictive. That forced all debounced elements to share a type
and munged together props relevant to separate concerns (debouncing and the
input).

This abstracts the Debouncing entirely into its own component. The new
DebounceComponent takes a single child, intercepts relevant props from it, wraps
them in a debounce, and reinjects them into the child. This should allow more
robust usage, for instance on a textarea as well as a number input. Additionally
it allows for more intuitive syntax to be written, as the input syntax doesn't
change from an un-debounced version.

For user ergonomics, the DebouncedComponent will also flush the debounce onBlur.