Create MaskedInput
In order to ensure data hygiene, it is useful to mask <input>s such that they do
not accept arbitrary user-entered data, rather it should be filtered to only
display when it meets certain criteria.

This is separate from data validation, which would require the comp. to be aware
of underlying data. Instead, this refers only to the shape of the data- is it
numeric, is it a string matching a certain pattern, etc.

As a proof of concept, and in order to better support the app's current usage,
this adds both the base component, MaskedInput, as well as a concrete impl.,
NumericInput, which masks out any input which is not a number, via a regex
replace